requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
efficientnet_pytorch==0.7.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.2.0
huggingface-hub==0.27.1
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.0.post0
lightning-utilities==0.11.9
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
monai==1.4.0
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
open3d==0.19.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
parso==0.8.4
phasepack==1.5
pillow==10.2.0
platformdirs==4.3.6
plotly==5.24.1
pretrainedmodels==0.7.4
prompt_toolkit==3.0.48
propcache==0.2.1
pure_eval==0.2.3
pyFFTW==0.15.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-image==0.25.0
scipy==1.15.0
segmentation_models_pytorch==0.4.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tifffile==2024.12.12
timm==1.0.13
torch==2.4.0+cu124
torchaudio==2.4.0+cu124
torchmetrics==1.6.1
torchvision==0.19.0+cu124
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
yarl==1.18.3
zipp==3.21.0
